feat(core): added ai functionality
