+++ eventcatalog/src/enterprise/README.md
@@ -1,0 +1,5 @@
+## EventCatalog License
+
+Usage of files in this directory and its subdirectories, and of EventCatalog Plugin / Enterprise features, is subject to the EventCatalog Commercial License, and conditional on having the right license from EventCatalog. Access to files in this directory and its subdirectories does not constitute permission to use this code or EventCatalog Enterprise Edition features.
+
+Unless otherwise noted, all files Copyright © 2025 EventCatalog Ltd.
