+++ eventcatalog/src/layouts/VerticalSideBarLayout.astro
@@ -5,7 +5,7 @@ interface Props {
   description?: string;
 }
 
-import { BookOpenText, Workflow, TableProperties, House, BookUser, MessageSquare } from 'lucide-react';
+import { BookOpenText, Workflow, TableProperties, House, BookUser, MessageSquare, BotMessageSquare } from 'lucide-react';
 import Header from '../components/Header.astro';
 import SEO from '../components/Seo.astro';
 import SideNav from '../components/SideNav/SideNav.astro';
@@ -94,8 +94,8 @@ const navigationItems = [
   },
   {
     id: '/chat',
-    label: 'Chat',
-    icon: MessageSquare,
+    label: 'AI Assistant',
+    icon: BotMessageSquare,
     href: buildUrl('/chat'),
     current: currentPath.includes('/chat'),
   },
