+++ eventcatalog/src/pages/chat/index.astro
@@ -1,13 +1,107 @@
 ---
 import VerticalSideBarLayout from '@layouts/VerticalSideBarLayout.astro';
-import Chat from '@components/Chat/Chat';
+import Chat from '@enterprise/ai-assistant/components/Chat';
+import config from '@config';
+import path from 'node:path';
+import fs from 'node:fs';
 
+import { Code } from 'astro-expressive-code/components';
+
+const isEnabled = config.chat?.enabled || false;
+
+const PROJECT_DIR = path.resolve(process.env.PROJECT_DIR || process.cwd());
+const GENERATED_AI_DIR = path.resolve(PROJECT_DIR, 'generated-ai');
+
+const directoryExists = fs.existsSync(GENERATED_AI_DIR);
+
+const generatorConfig = `
+ generators: [
+    [
+      "@eventcatalog/generator-ai", {
+        // If you want to chunk files into smaller chunks, set this to true
+        splitMarkdownFiles: true
+      }
+    ],
+  ],
+`;
 ---
 
 <VerticalSideBarLayout title="AI Chat">
   <div class="flex h-[calc(100vh-60px)] bg-white">
-    <!-- <Sidebar client:load /> -->
-      <Chat client:load />
+    {
+      isEnabled ? (
+        directoryExists ? (
+          <Chat client:only="react" catalogPath={PROJECT_DIR} />
+        ) : (
+          <div class="flex items-center justify-center w-full p-4 sm:p-8">
+            <div class="max-w-2xl text-center">
+              <h2 class="text-2xl font-bold text-gray-900 mb-4">EventCatalog AI Assistant Setup Required</h2>
+              <p class="text-gray-600 mb-6">
+                To use the AI Assistant, you need to generate the AI data first.
+                <br />
+                Please install the plugin and run the generator command to generate the AI data.
+              </p>
+
+              <div class="text-left space-y-4">
+                <div class="space-y-2">
+                  <h2 class="text-lg font-semibold  text-left">1. Install the generator</h2>
+                  <Code code={`npm i @eventcatalog/generator-ai`} lang="bash" frame="terminal" />
+                </div>
+
+                <div class="space-y-2">
+                  <h2 class="text-lg font-semibold text-left">2. Configure the generator</h2>
+                  <Code code={generatorConfig} lang="javascript" frame="terminal" title="eventcatalog.config.js" />
+                </div>
+
+                <div class="space-y-2">
+                  <h2 class="text-lg font-semibold  text-left">3. Generate the AI data</h2>
+                  <Code code={`npm run generate`} lang="bash" frame="none" />
+                </div>
+                <p class="text-gray-600 text-sm mt-2">
+                  After running the generator, a new folder called <code class="bg-gray-100 p-1 rounded-md">generated-ai</code>{' '}
+                  will be created in your project.
+                </p>
+
+                <p class="text-gray-600 text-sm mt-2">Once done, refresh this page.</p>
+              </div>
+            </div>
+          </div>
+        )
+      ) : (
+        <div class="flex items-center justify-center w-full p-4 sm:p-8">
+          <div class="max-w-6xl flex flex-col md:flex-row gap-8 md:gap-12 items-center">
+            {/* Left side - Illustration */}
+            <div class="flex-1">
+              <img
+                src="/images/ai-placeholder.png"
+                alt="AI Assistant Illustration"
+                class="w-full max-w-sm md:max-w-md mx-auto  rounded-lg shadow-lg"
+              />
+            </div>
+
+            {/* Right side - Content */}
+            <div class="flex-1">
+              <h2 class="text-2xl md:text-3xl font-bold text-gray-900 mb-4">EventCatalog AI Assistant</h2>
+              <p class="text-gray-600 mb-6 text-base md:text-lg">
+                Chat with your catalog to quickly find information, understand relationships, and get instant answers about your
+                architecture. Our AI-powered assistant makes exploring your event catalog faster and more intuitive than ever.
+              </p>
+
+              <div class="bg-gray-700 rounded-lg p-4 md:p-6 mb-4">
+                <h3 class="text-lg text-white font-semibold mb-3">Quick Setup</h3>
+                <Code code={`npm install @eventcatalog/generator-ai`} lang="bash" frame="none" />
+                <a href="https://www.eventcatalog.dev/ai-assistant" class="inline-flex items-center text-sm text-white mt-4">
+                  Learn more about setup →
+                </a>
+              </div>
+              <p class="text-gray-600 mb-6 text-sm">
+                Your data stays private and secure - all processing happens locally on your machine.
+              </p>
+            </div>
+          </div>
+        </div>
+      )
+    }
   </div>
 </VerticalSideBarLayout>
 
@@ -26,4 +120,3 @@ import Chat from '@components/Chat/Chat';
     scroll-behavior: smooth;
   }
 </style>
-
