+++ package.json
@@ -58,6 +58,7 @@
     "@langchain/core": "^0.3.40",
     "@mlc-ai/web-llm": "^0.2.78",
     "@parcel/watcher": "^2.4.1",
+    "@radix-ui/react-dialog": "^1.1.6",
     "@stoplight/json-schema-viewer": "^4.16.4",
     "@tailwindcss/typography": "^0.5.13",
     "@tanstack/react-table": "^8.17.3",
@@ -75,7 +76,6 @@
     "date-fns": "^4.1.0",
     "diff": "^7.0.0",
     "diff2html": "^3.4.48",
-    "faiss-node": "^0.5.1",
     "glob": "^10.4.1",
     "gray-matter": "^4.0.3",
     "html-to-image": "^1.11.11",
@@ -96,7 +96,6 @@
     "remark-directive": "^3.0.0",
     "remark-gfm": "^3.0.1",
     "rimraf": "^5.0.7",
-    "rxdb": "^16.6.1",
     "semver": "7.6.3",
     "shelljs": "^0.8.5",
     "tailwindcss": "^3.4.3",
